+++ pkgs/objective_c/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Use ffigen 18.2.0
 - `NSArray` is now a Dart `Iterable` and `NSMutableArray` is now a Dart `List`.
 - Add `.toNSNumber()` extension method to `int`, `double`, and `num`.
+- Add `DateTime.toNSDate()` and `NSDate.toDateTime()` extension methods.
 
 ## 7.1.0